Fix collection chapter
Add edit and status action for chapters in collection. if a draft
chapter “draft” is linked, it will be visualised in the list but it
will not be accessible (story api allow read access of drafts to staff
members or to authors only).
This doesn’t work for lists, and the LiteMentionSerializer used by the
CollectionSerializer points to all items by default.
We should modify the ListSerializer and add a custom to_reporesentation
method, as stated by
http://stackoverflow.com/questions/28163556/how-do-you-filter-a-nested-s
erializer-in-django-rest-framework