rais ValueError if nested function couldn't create proper empty dict
This happens when we have for instance:
```
data__category="string value"
data__category__en_US="string value"
```
since the nested function transforms "__"  paths into actual dict trees, and initialize every leave with 'None' default value. In this case `data['category']` is set to None, ready to receive a value (if any). In order to accomodate the en_US value in the following step, we assume that `data['category']` is a dict but we already have 'None' ...